[geometry] Split MeshcatParams into its own file
Generally we aim for one class per file for clarity, but for config
structs in particular we strongly prefer it so that config can use
the schema without pulling in the full class.
